Library

Types

Sad.Dingman — Type

Dingman river channel cross section.

Uses generalized expressions for cross-section geometry based on at-a-station hydraulic geometry relations.

Wb: bankfull width

Yb: bankfull depth

Ym: flow depth

r: geometry exponent

S0: bed slope

n: Manning roughness coefficient

See also: Rectangular

Sad.Rectangular — Type

Rectangular river channel cross section.

Wb: bankfull width

Yb: bankfull depth

Ym: flow depth

S0: bed slope

n: Manning roughness coefficient

See also: Dingman

Functions

Cross sections

Sad.width — Function

width(xs)

Calculate water surface width of cross section.

Arguments

xs: cross section

Sad.depth — Function

depth(xs)

Calculate average depth of cross section.

Arguments

xs: cross section

Priors

Sad.get_samples — Function

get_samples(p, samples)

Get quantiles from p according to weights provided by Latin Hypercube Sampling samples.

Sad.lhs_ensemble — Function

lhs_ensemble(nens, args...)

Generate an ensemble of size nens using Latin Hypercube Sampling of the list of distributions or collections provided as args.

Sad.prior_ensemble — Function

prior_ensemble(x::Vector{Float64}, Qp::Distribution, np::Distribution, rp::Distribution, zp::Distribution, nens::Int)

Generate a prior ensemble of discharge, roughness coefficient, channel shape parameter, and downstream bed elevation from provided distributions or sample data.

Arguments

x: distances between cross sections

Qp: discharge prior distribution

np: roughness coefficient distribution

rp: channel shape parameter distribution

zp: downstream bathymetry distribution

nens: ensemble size

Sad.gvf_ensemble! — Function

gvf_ensemble!(H::Vector{Float64}, S, x::Vector{Float64}, hbf::Vector{Float64}, wbf::Vector{Float64}, Qe::Vector{Float64}, ne::Vector{Float64}, re::Vector{Float64}, ze)

Generate an ensemble of water height profiles from Gradually-Varied-Flow simulations and associated profiles of bed elevation.

H: water surface elevation

S: bed slope

x: channel chainage

hbf: bankfull water surface elevation

wbf: bankfull width

Qe: ensemble discharge

ne: ensemble roughness coefficient

re: ensemble channel shape parameter

ze: ensemble bed elevation profiles

Sad.rejection_sampling — Function

rejection_sampling(Qp, np, rp, zp, x, H, S0, hbc, wbf, hbf, nens)

Use rejection sampling to select a subset of the prior ensemble.

Arguments

Qp: prior discharge distribution

np: prior roughness coefficient distribution

rp: prior channel shape parameter distribution

zp: prior distribution for downstream bed elevation

x: channel chainage

H: time series of observed water surface elevation profiles

S0: prior estimate of channel bed slope

hbc: mean downstream flow depth used as boundary condition

wbf: bankfull width

hbf: bankfull water surface elevation

nens: ensemble size

Sad.priors — Function

priors(ncfile)

Derive prior distributions for discharge, roughness coefficient, channel shape parameter, and bed elevation by reading the SWORD a-priori database.

Arguments

ncfile: NetCDF file of SWORD database

priors(qwbm, hmin)

Derive distributions for discharge, roughness coefficient, channel shape parameter, and bed elevation using uninformative priors.

Arguments

qwbm: mean discharge

hmin: minimum downstream water surface elevation

Kalman filters

Sad.letkf — Function

letkf(A::Matrix{Float64}, d::Vector{Float64}, HA::Matrix{Float64}, E::Matrix{Float64}, ρ::Float64, diagR::Bool)

Apply the Local Ensemble Transform Kalman Filter algorithm.

Arguments

A: state variable ensemble matrix

d: observation vector

HA: model-predicted observation ensemble matrix

E: observation error ensemble matrix

ρ: covariance inflation parameter (optional, default value of 1.05)

diagR: force observation error covariance to be diagonal, i.e. no spatial correlation (optional, default is false)

Gradually Varied Flow

Sad.froude — Function

froude(Q::Float64, xs::CrossSection)

Calculate the Froude number.

Arguments

Q: discharge

xs: cross section

Sad.dydx — Function

dydx(y, p, x)

Ordinary differential equation describing the Gradually-Varied-Flow model.

Sad.gvf — Function

Calculate a water surface profile by solving the Gradually-Varied-Flow equation.

Q: discharge

ybc: downstream boundary condition for depth

S0: bed slope for reach

n: roughness coefficient

x: downstream distance for each cross section

wbf: bankfull width for each cross section

ybf: bankfull depth for each cross section

r: channel geometry coefficient for each cross section

Sad

Sad.assimilate — Function

assimilate(H::Vector{Float64})

Assimilate SWOT observations for river reach.

Arguments

H: water surface elevation

W: water surface width

x: downstream distance for each cross section

wbf: bankfull width

hbf: bankfull depth

Qₚ: prior probability distribution for discharge

nₚ: prior probability distribution for roughness coefficient

rₚ: prior probability distribution for channel shape parameter

zₚ: prior distribution for downstream bed elevation

nens: ensemble size

ϵₒ: observation error standard deviation

Sad.bathymetry! — Function

bathymetry!(ze, Se, Qe, ne, re, x, hbf, wbf, H, ϵ₀::Float64=0.01)

Estimate channel bed slope and thalweg (elevation) by assimilating SWOT water surface elevation observations.

Arguments

ze: ensemble of bed elevations

Se: ensemble of bed slopes

Qe: ensemble of discharge

ne: ensemble of roughness coefficient

re: ensemble of channel shape parameters

x: channel chainage

hbf: bankfull water surface elevation

wbf: bankfull width

H: time-averaged water surface elevation observed

ϵ₀: observation error standard deviation